% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_impute.R
\name{did_impute}
\alias{did_impute}
\title{Difference-in-differences imputation estimator (BJS 2024)}
\usage{
did_impute(
  df,
  y,
  i,
  t,
  Ei,
  controls = character(),
  fe = NULL,
  timecontrols = character(),
  aw = NULL,
  unitcontrols = character(),
  wtr = character(),
  sum = FALSE,
  horizons = NULL,
  allhorizons = FALSE,
  hbalance = FALSE,
  hetby = character(),
  project = character(),
  minn = 30,
  saveweights = FALSE,
  shift = 0,
  pretrends = 0,
  cluster = "",
  avgeffectsby = character(),
  leaveoneout = FALSE,
  nose = FALSE,
  delta = NULL,
  seed = 1L
)
}
\description{
Difference-in-differences imputation estimator (BJS 2024)
}
