% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didimpute-package.R
\docType{package}
\name{didimpute-package}
\alias{didimpute}
\alias{didimpute-package}
\title{didimpute: Difference-in-Differences Imputation Estimator (Borusyak, Jaravel & Spiess)}
\description{
A faithful R port of the Python did_imputation package implementing the imputation estimator of Borusyak, Jaravel and Spiess (2024) for staggered difference-in-differences, with event-study aggregation and plotting.
}
\author{
\strong{Maintainer}: Xao \email{xao@hbs.edu}

Authors:
\itemize{
  \item Xao \email{xao@hbs.edu}
}

}
\keyword{internal}
